DFS - 6192, 6193, 6217 and 6202 (#393)
* DFS-6193 and DFS-6202

DFS-6193:Fixed the scroll bar issue on Add subfolder modal, DFS-6202:Fixed the issue with the CLM Files widget not working when parent folder is created as part of the object mapping

* Fixed DFS-6217

Fixed issue with the onboarding checklist checkboxes not working post Step 1

* Fixed variable declaration format

* Fixed DFS-6192

Object folder name and path fields in object mapping table are now text fields and not hyperlinks. This should prevent navigation to broken links on click of these field values.

* Removed obsolete code

Removed tooltip attributes from the code. They're not required for a text type datatable column field

* Removed hardcoded strings and added custom labels

* PR Review: Fixed formatting

* Fixed formatting issues: spaces around operators