CLM - UI fixes (#441)
* Fixed issue of extended whitespace at bottom of the page

Showing a tree of records that exceeded a specific height caused the page to show extended whitespace at the bottom. Enforced a max height to contain it.

* Fixed DFS-6290: Help card UI changes

* Fixed incorrect button label on Integrations page help card

* Removed CLMCardModel aura cmp and its event

The aura component and its namesake event are not required anymore and can be safely deleted. References to the CLMCardModelEvent from CLMIntegrationLayout have been removed.

* Removed stale references to CLMCardModel aura component

* Fixed: DFS-6292-illustration to show up if no objects are mapped

* Removed stale reference to CLMCardModel

* DFS-6266:Changed image on Help FAQ page

* Removed inline css styles

* Adjusted img position and size

* Removed CLMTroubleshoot aura cmp  and CLMHavingTrouble static resource

* Addressed PR changes

Returned a list as opposed to a map from the apex controller. Also, replaced Object.values with isEmpty method of the Aura global variable. Additionally, refactored the init method to skip a server call if there are no mapped objects.